Move font header files to same folder as sketch
Arduino IDE does not save the directory that includes
the header files when saving it in a non read-only
location. Let's work around this by placing the headers
in the same directory as the sketch so they can be
picked up both when editing/viewing the sketch but also
when saving it locally for further modifications.

Fixes: #19